[config] Add config in core
What I did:
- add config module in core
- move CudaConfig from layer.hpp to config.hpp
