+++ src/core/CMakeLists.txt
@@ -1,4 +1,5 @@
 # Add all subdirectories
 add_subdirectory(type)
 add_subdirectory(memory)
-add_subdirectory(status)  # Add status library
+add_subdirectory(status)
+add_subdirectory(config)
