+++ src/core/config/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(config INTERFACE)
+target_include_directories(config INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
